backend/requirements.txt
fastapi
uvicorn
python-jose
requests
python-dotenv
sqlalchemy
alembic
psycopg
psycopg2